requirements.txt
python-telegram-bot==20.6
schedule
nest_asyncio
